requirements.txt
dj-database-url==0.5.0
Django==2.0.5
djangorestframework==3.8.2
pbr==4.0.3
pytz==2018.4
six==1.11.0
stevedore==1.28.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
